pyproject.toml
[project]
name = "mqtt-to-stuff"
version = "0.1.0"
description = ""
authors = [
    {name = "leprechaun",email = "leprechaun@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "pytest (>=8.3.5,<9.0.0)",
    "paho-mqtt (>=2.1.0,<3.0.0)",
    "prometheus-client (>=0.21.1,<0.22.0)",
    "deltalake (>=0.25.5,<0.26.0)",
    "polars (>=1.27.1,<2.0.0)"
#    "marimo (>=0.13.4,<0.14.0)",
#    "altair (>=5.5.0,<6.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
